Split code; Use empty repo type as default
