+++ src/Controller/OrganizationController.php
@@ -4,22 +4,15 @@ declare(strict_types=1);
 
 namespace Buddy\Repman\Controller;
 
-use Buddy\Repman\Entity\Organization\Package\Metadata;
 use Buddy\Repman\Entity\User;
-use Buddy\Repman\Entity\User\OAuthToken;
-use Buddy\Repman\Form\Type\Organization\AddPackageType;
 use Buddy\Repman\Form\Type\Organization\ChangeAliasType;
 use Buddy\Repman\Form\Type\Organization\ChangeNameType;
 use Buddy\Repman\Form\Type\Organization\CreateType;
 use Buddy\Repman\Form\Type\Organization\GenerateTokenType;
-use Buddy\Repman\Message\Organization\AddPackage;
 use Buddy\Repman\Message\Organization\ChangeAlias;
 use Buddy\Repman\Message\Organization\ChangeName;
 use Buddy\Repman\Message\Organization\CreateOrganization;
 use Buddy\Repman\Message\Organization\GenerateToken;
-use Buddy\Repman\Message\Organization\Package\AddBitbucketHook;
-use Buddy\Repman\Message\Organization\Package\AddGitHubHook;
-use Buddy\Repman\Message\Organization\Package\AddGitLabHook;
 use Buddy\Repman\Message\Organization\Package\RemoveBitbucketHook;
 use Buddy\Repman\Message\Organization\Package\RemoveGitHubHook;
 use Buddy\Repman\Message\Organization\Package\RemoveGitLabHook;
@@ -32,21 +25,12 @@ use Buddy\Repman\Query\User\Model\Organization;
 use Buddy\Repman\Query\User\Model\Package;
 use Buddy\Repman\Query\User\OrganizationQuery;
 use Buddy\Repman\Query\User\PackageQuery;
-use Buddy\Repman\Service\BitbucketApi;
-use Buddy\Repman\Service\GitHubApi;
-use Buddy\Repman\Service\GitLabApi;
 use Buddy\Repman\Service\Organization\AliasGenerator;
 use Ramsey\Uuid\Uuid;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Validator\Constraints\NotBlank;
 
 final class OrganizationController extends AbstractController
 {
@@ -68,7 +52,6 @@ final class OrganizationController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            /** @var User */
             $user = $this->getUser();
 
             $this->dispatchMessage(new CreateOrganization(
@@ -114,78 +97,6 @@ final class OrganizationController extends AbstractController
             'count' => $count,
             'organization' => $organization,
         ]);
-    }
-
-    /**
-     * @Route("/organization/{organization}/package/new/{type?}", name="organization_package_new", methods={"GET","POST"}, requirements={"organization"="%organization_pattern%"})
-     */
-    public function packageNew(Organization $organization, Request $request, GithubApi $githubApi, GitlabApi $gitlabApi, BitbucketApi $bitbucketApi, ?string $type): Response
-    {
-        $type ??= 'git';
-        $form = $this->createForm(AddPackageType::class);
-        $form->get('formUrl')->setData($this->generateUrl(
-            $request->attributes->get('_route'),
-            ['organization' => $organization->alias()],
-            RouterInterface::ABSOLUTE_URL
-        ));
-        $form->get('type')->setData($type);
-
-        switch ($type) {
-            case 'git':
-            case 'mercurial':
-            case 'subversion':
-            case 'pear':
-                $response = $this->packageNewFromUrl('url', $form, $organization, $request);
-                break;
-            case 'path':
-            case 'artifact':
-                $response = $this->packageNewFromUrl($type, $form, $organization, $request);
-                break;
-            case 'github':
-                $response = $this->packageNewFromGitHub($form, $organization, $request, $githubApi);
-                break;
-            case 'gitlab':
-                $response = $this->packageNewFromGitLab($form, $organization, $request, $gitlabApi);
-                break;
-            case 'bitbucket':
-                $response = $this->packageNewFromBitbucket($form, $organization, $request, $bitbucketApi);
-                break;
-            default:
-                throw new NotFoundHttpException();
-        }
-
-        if ($response instanceof Response) {
-            return $response;
-        }
-
-        return $this->render('organization/addPackage.html.twig', [
-            'organization' => $organization,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @param string[] $choices
-     *
-     * @return array<string|array>
-     */
-    public function repositoriesChoiceType(array $choices): array
-    {
-        return [
-            'repositories',
-            ChoiceType::class, [
-                'choices' => $choices,
-                'label' => false,
-                'expanded' => false,
-                'multiple' => true,
-                'attr' => [
-                    'class' => 'form-control selectpicker',
-                    'data-live-search' => 'true',
-                    'data-style' => 'btn-secondary',
-                    'title' => 'select repository',
-                ],
-            ],
-        ];
     }
 
     /**
@@ -376,132 +287,5 @@ final class OrganizationController extends AbstractController
         $user = parent::getUser();
 
         return $user;
-    }
-
-    private function packageHasBeenAdded(Organization $organization): Response
-    {
-        $this->addFlash('success', 'Packages has been added and will be synchronized in the background');
-
-        return $this->redirectToRoute('organization_packages', ['organization' => $organization->alias()]);
-    }
-
-    private function packageNewFromUrl(string $label, FormInterface $form, Organization $organization, Request $request): ?Response
-    {
-        $form->add(
-            'url',
-            TextType::class,
-            [
-                'constraints' => [new NotBlank()],
-                'label' => $label,
-            ],
-        );
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $type = $form->get('type')->getData();
-
-            $this->dispatchMessage(new AddPackage(
-                $id = Uuid::uuid4()->toString(),
-                $organization->id(),
-                $form->get('url')->getData(),
-                in_array($type, ['git', 'mercurial', 'subversion'], true) ? 'vcs' : $type
-            ));
-            $this->dispatchMessage(new SynchronizePackage($id));
-
-            return $this->packageHasBeenAdded($organization);
-        }
-
-        return null;
-    }
-
-    private function packageNewFromGitHub(FormInterface $form, Organization $organization, Request $request, GithubApi $api): ?Response
-    {
-        $token = $this->getUser()->oauthToken(OAuthToken::TYPE_GITHUB);
-        if ($token === null) {
-            return $this->redirectToRoute('fetch_github_package_token', ['organization' => $organization->alias()]);
-        }
-
-        $repos = $api->repositories($token->accessToken());
-        $choices = array_combine($repos, $repos);
-        $form->add(...$this->repositoriesChoiceType(is_array($choices) ? $choices : []));
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            foreach ($form->get('repositories')->getData() as $repo) {
-                $this->dispatchMessage(new AddPackage(
-                    $id = Uuid::uuid4()->toString(),
-                    $organization->id(),
-                    "https://github.com/{$repo}",
-                    'github-oauth',
-                    [Metadata::GITHUB_REPO_NAME => $repo]
-                ));
-                $this->dispatchMessage(new SynchronizePackage($id));
-                $this->dispatchMessage(new AddGitHubHook($id));
-            }
-
-            return $this->packageHasBeenAdded($organization);
-        }
-
-        return null;
-    }
-
-    private function packageNewFromGitLab(FormInterface $form, Organization $organization, Request $request, GitlabApi $api): ?Response
-    {
-        $token = $this->getUser()->oauthToken(OAuthToken::TYPE_GITLAB);
-        if ($token === null) {
-            return $this->redirectToRoute('fetch_gitlab_package_token', ['organization' => $organization->alias()]);
-        }
-
-        $projects = $api->projects($token->accessToken());
-        $form->add(...$this->repositoriesChoiceType(array_flip($projects->names())));
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            foreach ($form->get('repositories')->getData() as $projectId) {
-                $this->dispatchMessage(new AddPackage(
-                    $id = Uuid::uuid4()->toString(),
-                    $organization->id(),
-                    $projects->get($projectId)->url(),
-                    'gitlab-oauth',
-                    [Metadata::GITLAB_PROJECT_ID => $projectId]
-                ));
-                $this->dispatchMessage(new SynchronizePackage($id));
-                $this->dispatchMessage(new AddGitLabHook($id));
-            }
-
-            return $this->packageHasBeenAdded($organization);
-        }
-
-        return null;
-    }
-
-    private function packageNewFromBitbucket(FormInterface $form, Organization $organization, Request $request, BitbucketApi $api): ?Response
-    {
-        $token = $this->getUser()->oauthToken(OAuthToken::TYPE_BITBUCKET);
-        if ($token === null) {
-            return $this->redirectToRoute('fetch_bitbucket_package_token', ['organization' => $organization->alias()]);
-        }
-
-        $repos = $api->repositories($token->accessToken());
-        $form->add(...$this->repositoriesChoiceType(array_flip($repos->names())));
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            foreach ($form->get('repositories')->getData() as $repoUuid) {
-                $this->dispatchMessage(new AddPackage(
-                    $id = Uuid::uuid4()->toString(),
-                    $organization->id(),
-                    $repos->get($repoUuid)->url(),
-                    'bitbucket-oauth',
-                    [Metadata::BITBUCKET_REPO_NAME => $repos->get($repoUuid)->name()]
-                ));
-                $this->dispatchMessage(new SynchronizePackage($id));
-                $this->dispatchMessage(new AddBitbucketHook($id));
-            }
-
-            return $this->packageHasBeenAdded($organization);
-        }
-
-        return null;
     }
 }
